glacier-backend/ml/requirements.txt
numpy>=1.24.0
onnxruntime>=1.16.0
opencv-python>=4.8.0
pillow>=10.0.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
